switch ram model to 64 bit

+++ examples/ex00/BDPI/image_loader.h
@@ -18,12 +18,12 @@ public:
     uint8_t get_pixel_at(uint32_t index);
 
     //writes each value it gets from data buffer to a new address in memory
-    void write_to_mem_seq(bluesy::ptr_type ram_ptr, uint32_t start_addr = 0x0);
+    void write_to_mem_seq(bluesy::ptr_type ram_ptr, uint64_t start_addr = 0x0);
 
     //interleaves RGB values of pixels so that each word contains all three values for a pixel
-    void write_to_mem_interleaved(bluesy::ptr_type ram_ptr, uint32_t start_addr = 0x0);
+    void write_to_mem_interleaved(bluesy::ptr_type ram_ptr, uint64_t start_addr = 0x0);
 
-    void write_to_mem_gray_packed(bluesy::ptr_type ram_ptr, uint32_t start_addr = 0x0);
+    void write_to_mem_gray_packed(bluesy::ptr_type ram_ptr, uint64_t start_addr = 0x0);
 
 private:
     CImg<uint8_t> m_image;
